Run tests on tecnai – success, new interface works!
